+++ testsutils/authenticated_test_service.py
@@ -90,6 +90,80 @@ def swagger():
                                    }
                                ]
                            }
+                       },
+                       '/test/basic/authentication/success': {
+                           'get': {
+                               'operationId': 'get_test_basic_authentication_success',
+                               'responses': {
+                                   '200': {
+                                       'description': 'return value'
+                                   }
+                               },
+                               'security': [
+                                   {
+                                       'basic_auth_success': [
+                                       ]
+                                   }
+                               ]
+                           }
+                       },
+                       '/test/basic/and/api/key/authentication/success': {
+                           'get': {
+                               'operationId': 'get_test_basic_and_api_key_authentication_success',
+                               'responses': {
+                                   '200': {
+                                       'description': 'return value'
+                                   }
+                               },
+                               'security': [
+                                   {
+                                       'basic_auth_success': [
+                                       ],
+                                       'api_key_auth_success': [
+                                       ]
+                                   }
+                               ]
+                           }
+                       },
+                       '/test/basic/or/api/key/authentication/success': {
+                           'get': {
+                               'operationId': 'get_test_basic_or_api_key_authentication_success',
+                               'responses': {
+                                   '200': {
+                                       'description': 'return value'
+                                   }
+                               },
+                               'security': [
+                                   {
+                                       'basic_auth_success': [
+                                       ]
+                                   },
+                                   {
+                                       'api_key_auth_success': [
+                                       ]
+                                   }
+                               ]
+                           }
+                       },
+                       '/test/api/key/or/basic/authentication/success': {
+                           'get': {
+                               'operationId': 'get_test_api_key_or_basic_authentication_success',
+                               'responses': {
+                                   '200': {
+                                       'description': 'return value'
+                                   }
+                               },
+                               'security': [
+                                   {
+                                       'api_key_auth_success': [
+                                       ]
+                                   },
+                                   {
+                                       'basic_auth_success': [
+                                       ]
+                                   }
+                               ]
+                           }
                        }
                    },
                    securityDefinitions={
@@ -130,6 +204,9 @@ def swagger():
                            "type": "apiKey",
                            "in": "header",
                            "name": "X-API-KEY"
+                       },
+                       'basic_auth_success': {
+                           "type": "basic"
                        }
                    })
 
@@ -157,6 +234,41 @@ def get_test_oauth2_authentication_success_quick_expiry():
 @app.route('/test/api/key/authentication/success', methods=['GET'])
 def get_test_api_key_authentication_success():
     return jsonify([{'X-API-KEY': request.headers.get('X-API-KEY')}])
+
+
+@app.route('/test/basic/authentication/success', methods=['GET'])
+def get_test_basic_authentication_success():
+    return jsonify([{'Authorization': request.headers.get('Authorization')}])
+
+
+@app.route('/test/basic/and/api/key/authentication/success', methods=['GET'])
+def get_test_basic_and_api_key_authentication_success():
+    return jsonify([
+        {
+            'Authorization': request.headers.get('Authorization'),
+            'X-API-KEY': request.headers.get('X-API-KEY')
+        }
+    ])
+
+
+@app.route('/test/basic/or/api/key/authentication/success', methods=['GET'])
+def get_test_basic_or_api_key_authentication_success():
+    return jsonify([
+        {
+            'Authorization': request.headers.get('Authorization'),
+            'X-API-KEY': request.headers.get('X-API-KEY')
+        }
+    ])
+
+
+@app.route('/test/api/key/or/basic/authentication/success', methods=['GET'])
+def get_test_api_key_or_basic_authentication_success():
+    return jsonify([
+        {
+            'Authorization': request.headers.get('Authorization'),
+            'X-API-KEY': request.headers.get('X-API-KEY')
+        }
+    ])
 
 
 def start_server(port):