+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Enhancements ###
 
 - API Key Security definition support. You should specify your API Key thanks to api_key keyword in security_details property.
+- Basic security definition support. You should specify your username and password thanks to username and password keywords in security_details property.
 - Slight performance improvement on UDFs loading.
 - If more than one supported authentication step is required, all steps are now performed.